requirements.txt
altgraph==0.17.5
babel==2.18.0
customtkinter==5.2.2
darkdetect==0.8.0
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
pygame==2.6.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
pystray==0.19.5
pywin32-ctypes==0.2.3
six==1.17.0
tkcalendar==1.6.1
typing_extensions==4.15.0
Windows-Toasts==1.3.1
winotify==1.1.0
winrt-runtime==3.2.1
winrt-Windows.Data.Xml.Dom==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.UI.Notifications==3.2.1
winsdk==1.0.0b10
